pyproject.toml
[project]
name = "adaptive-study-agent"
version = "1.0.0"
description = "A single-agent self-directed learning system built with LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-chroma>=0.2.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
